Drunken Brawlening: Intoxication alters your unarmed effectiveness, bartenders are stronger fighters while drunk, deterministic stagger combo is back (your average assistant doesn't get knockdowns) (#85449)
## About The Pull Request

### Intoxication and Punching

Being drunk juuuust enough allows you to land stronger punches and take
punches a little more easily.

However, being TOO drunk makes you WORSE at these things, and makes your
unarmed accuracy an unarmed PENALTY. Don't get drunk if you can't hold
your liquor.

However, fighting drunk makes you really sick, so be careful as to not
suddenly start vomiting everywhere.

### Drunken Brawler

Bartenders are now drunken brawlers. They gain additional punching
prowess if they're ever intoxicated, it does not matter how much nor do
they ever suffer a penalty as a result of being TOO drunk.

The more damage the bartender has taken, the stronger they hit back.
(added by recommendation from WalterTruck) I have to mess the scaling on
this more.

### Stagger Combo/Unarmed Combos

The stagger combo is back, but significantly less stun heavy except on
the extreme ends of unarmed upgrading.

When your opponent has been sufficiently damaged enough, staggered and
doesn't have any riot protection gear on, you can inflict additional
status effects when you punch an opponent.

The necessary value for this effect is: Half the target's Brute and Burn
damage combined exceeds a value of: Limb accuracy + armor block (min.
40. max 200)

While the thresholds are more deterministic, the effects are random. The
better you are at punching, the better the effect, but the higher your
targets armor, the weaker the effect. All effects are positive for the
attacker and negative for the target.

The dice roll is such; (a roll between -20 to 20) + limb accuracy -
armor block

| Stage/Value | Effect |
| ------------- | ------------- |
| Stage 1 (-Infinity to 0  | Staggered for 1 second |
| Stage 2 (1 to 10)  | Eye Blur for 5 seconds |
| Stage 3 (11 to 30)  | Eye Blur, Dizzying, Confusion for 5 seconds |
| Stage 4 (31 to 40)  | Blindness, Dizzying, Confusion for 5 seconds |
| Stage 5 (41 to 45)  | Knockdown for 4 seconds |
| Stage 6 (46 to Infinity) | Knockdown for 4 seconds, 5 Brute with huge
wound chance |

Stages 5 and 6 are not possible to reach for roundstart humans who
aren't intoxicated. It is almost impossible to get high effects against
someone who is in full security gear as a roundstart human

### Minor Changes

Any unarmed related thresholds or values that would use Stamina damage
now use half of the total of Burn and Brute damage on the relevant
individual. This affects punch accuracy, the threshold for getting a
stagger combo, and the threshold for someone being more vulnerable to
grabs.

## Why It's Good For The Game

I just really wanted to add some drunken brawling/drunk fist style
martial art stuff because I thought it would be fun, and bar brawls are
where most people start fist fights. It's thematically entertaining.

The bartender, by extension, should be pretty proficient at this kind of
brawling. It also works out great when working with chefs to beat people
up with their fists in their own, more unrefined style.

The situation with the unarmed knockdown/stamina pr was kind of a mess,
since at the time I was already going to take it out after discussing
things with Walter and we were working out some alternatives together.
But uh...well, it wasn't to be. Particularly, there was a lot of stuff
in unarmed fighting that used Stamina damage that didn't actually get
removed at all. So this cleans that up.

I said I'd find a good alternative down the line to fill that gap, and
this is it. For your run-of-the-mill fistfights, there are no random
knockdowns. It's only when you are actually trying to brawl with a
cybernetically enhanced human that this kind of thing starts to crop up.
Or your opponent is a golem. Or you are both really drunk (and thus
everyone starts vomiting because they're getting beaten up)

The average assistant beating up another assistant is at best going to
inflict some eye blurring and dizziness, assuming the fight even lasts
that long. But this should be sufficient to create space to escape, or
even meaningfully disable someone who might simply be invading your
department. It is much less helpful against a more sufficiently geared
opponent.

The main reason for focusing on this is to give the average joe some
kind of way for a heel turn in the event they're having to scuffle with
departmental invaders or escape some determined opponent. It does not,
however, allow them to manhandle a nuclear operative in an elite suit
while our would-be hero is entirely butt-naked and drunk. Not that they
shouldn't try, that's funny as hell.

Thus, the effects chosen are more useful for escaping and befuddlement
than necessarily stopping the fight outright. If you have improved arms,
your arms are weapons and start to act like them, so you start seeing
more significant results.

It is also a nice thematic tie to unarmed's zany cousin; tackling, since
the randomized effects are borrowing the logic form there, and it is
cool to have those tied a bit closer.

## Changelog
:cl:
add: Drunken fist fighting now has bonuses and penalties based on how
intoxicated you are. Controlled liquor intake could make you a better
brawler. Though you might vomit if you go too hard.
add: Bartenders are now Drunken Brawlers. If they're drunk at all (no
matter how drunk), they're stronger at fist fighting.
add: Reintroduces the deterministic stagger combo. The effects of the
combo are more varied, and based on attacker unarmed effectiveness and
the defender's armor. Read the PR for more details. It's complicated.
qol: Unarmed effects that would utilize stamina values now use a split
of half brute, half burn to determine outcomes or for meeting
thresholds. This affects; punch accuracy, stagger combo thresholds and
grab vulnerability thresholds.
/:cl: